Move method to test, so that it don't need to be tested
