+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/heuristic/selector/move/generic/list/kopt/KOptListMove.java
@@ -104,92 +104,6 @@ public class KOptListMove<Solution_> extends AbstractMove<Solution_> {
                 .collect(Collectors.toSet());
     }
 
-    /**
-     * Create a sequential or non-sequential k-opt from the supplied pairs of undirected removed and added edges.
-     *
-     * @param listVariableDescriptor
-     * @param indexVariableSupply
-     * @param entity The entity
-     * @param removedEdgeList The edges to remove. For each pair {@code (edgePairs[2*i], edgePairs[2*i+1])},
-     *        it must be the case {@code edgePairs[2*i+1]} is either the successor or predecessor of
-     *        {@code edgePairs[2*i]}. Additionally, each edge must belong to the given entity's
-     *        list variable.
-     * @param addedEdgeList The edges to add. Must contain only endpoints specified in the removedEdgeList.
-     * @return A new sequential or non-sequential k-opt move with the specified undirected edges removed and added.
-     * @param <Solution_>
-     */
-    public static <Solution_> KOptListMove<Solution_> fromRemovedAndAddedEdges(
-            ListVariableDescriptor<Solution_> listVariableDescriptor,
-            IndexVariableSupply indexVariableSupply,
-            Object entity,
-            List<Object> removedEdgeList,
-            List<Object> addedEdgeList) {
-
-        if (addedEdgeList.size() != removedEdgeList.size()) {
-            throw new IllegalArgumentException(
-                    "addedEdgeList (" + addedEdgeList + ") and removedEdgeList (" + removedEdgeList + ") have the same size");
-        }
-
-        if ((addedEdgeList.size() % 2) != 0) {
-            throw new IllegalArgumentException(
-                    "addedEdgeList and removedEdgeList are invalid: there is an odd number of endpoints.");
-        }
-
-        if (!addedEdgeList.containsAll(removedEdgeList)) {
-            throw new IllegalArgumentException("addedEdgeList (" + addedEdgeList + ") is invalid; it contains endpoints "
-                    + "that are not included in the removedEdgeList (" + removedEdgeList + ").");
-        }
-
-        Function<Object, Object> successorFunction =
-                getSuccessorFunction(listVariableDescriptor, ignored -> entity, indexVariableSupply);
-
-        for (int i = 0; i < removedEdgeList.size(); i += 2) {
-            if (successorFunction.apply(removedEdgeList.get(i)) != removedEdgeList.get(i + 1)
-                    && successorFunction.apply(removedEdgeList.get(i + 1)) != removedEdgeList.get(i)) {
-                throw new IllegalArgumentException("removedEdgeList (" + removedEdgeList + ") contains an invalid edge ((" +
-                        removedEdgeList.get(i) + ", " + removedEdgeList.get(i + 1) + ")).");
-            }
-        }
-
-        Object[] tourArray = new Object[removedEdgeList.size() + 1];
-        Integer[] incl = new Integer[removedEdgeList.size() + 1];
-        for (int i = 0; i < removedEdgeList.size(); i += 2) {
-            tourArray[i + 1] = removedEdgeList.get(i);
-            tourArray[i + 2] = removedEdgeList.get(i + 1);
-            int addedEdgeIndex = identityIndexOf(addedEdgeList, removedEdgeList.get(i));
-
-            if (addedEdgeIndex % 2 == 0) {
-                incl[i + 1] = identityIndexOf(removedEdgeList, addedEdgeList.get(addedEdgeIndex + 1)) + 1;
-            } else {
-                incl[i + 1] = identityIndexOf(removedEdgeList, addedEdgeList.get(addedEdgeIndex - 1)) + 1;
-            }
-
-            addedEdgeIndex = identityIndexOf(addedEdgeList, removedEdgeList.get(i + 1));
-            if (addedEdgeIndex % 2 == 0) {
-                incl[i + 2] = identityIndexOf(removedEdgeList, addedEdgeList.get(addedEdgeIndex + 1)) + 1;
-            } else {
-                incl[i + 2] = identityIndexOf(removedEdgeList, addedEdgeList.get(addedEdgeIndex - 1)) + 1;
-            }
-        }
-
-        KOptDescriptor<Solution_> descriptor = new KOptDescriptor<>(tourArray,
-                incl,
-                getSuccessorFunction(listVariableDescriptor,
-                        ignored -> entity,
-                        indexVariableSupply),
-                getBetweenPredicate(indexVariableSupply));
-        return descriptor.getKOptListMove(listVariableDescriptor, indexVariableSupply, entity);
-    }
-
-    private static int identityIndexOf(List<Object> sourceList, Object query) {
-        for (int i = 0; i < sourceList.size(); i++) {
-            if (sourceList.get(i) == query) {
-                return i;
-            }
-        }
-        return -1;
-    }
-
     public static <Solution_> Function<Object, Object> getSuccessorFunction(
             ListVariableDescriptor<Solution_> listVariableDescriptor,
             SingletonInverseVariableSupply inverseVariableSupply,
