+++ core/optaplanner-core-impl/src/test/java/org/optaplanner/core/impl/heuristic/selector/move/generic/list/kopt/KOptListMoveTest.java
@@ -1,9 +1,12 @@
 package org.optaplanner.core.impl.heuristic.selector.move.generic.list.kopt;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.optaplanner.core.impl.heuristic.selector.move.generic.list.kopt.KOptListMove.getBetweenPredicate;
+import static org.optaplanner.core.impl.heuristic.selector.move.generic.list.kopt.KOptListMove.getSuccessorFunction;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Function;
 
 import org.junit.jupiter.api.Test;
 import org.optaplanner.core.impl.domain.solution.descriptor.SolutionDescriptor;
@@ -41,7 +44,7 @@ class KOptListMoveTest {
         IndexVariableListener indexVariableListener = (IndexVariableListener) indexVariableSupply;
         TestdataListEntity e1 = new TestdataListEntity("e1", new ArrayList<>(List.of(v1, v2, v3, v4, v5, v6)));
         indexVariableListener.afterListVariableChanged(scoreDirector, e1, 0, 6);
-        KOptListMove<TestdataListSolution> kOptListMove = KOptListMove.fromRemovedAndAddedEdges(variableDescriptor,
+        KOptListMove<TestdataListSolution> kOptListMove = fromRemovedAndAddedEdges(variableDescriptor,
                 indexVariableSupply,
                 e1,
                 List.of(v6, v1,
@@ -68,7 +71,7 @@ class KOptListMoveTest {
 
         indexVariableListener.afterListVariableChanged(scoreDirector, e1, 0, 8);
 
-        KOptListMove<TestdataListSolution> kOptListMove = KOptListMove.fromRemovedAndAddedEdges(variableDescriptor,
+        KOptListMove<TestdataListSolution> kOptListMove = fromRemovedAndAddedEdges(variableDescriptor,
                 indexVariableSupply,
                 e1,
                 List.of(
@@ -102,7 +105,7 @@ class KOptListMoveTest {
 
         indexVariableListener.afterListVariableChanged(scoreDirector, e1, 0, 8);
 
-        KOptListMove<TestdataListSolution> kOptListMove = KOptListMove.fromRemovedAndAddedEdges(variableDescriptor,
+        KOptListMove<TestdataListSolution> kOptListMove = fromRemovedAndAddedEdges(variableDescriptor,
                 indexVariableSupply,
                 e1,
                 List.of(v8, v1,
@@ -135,7 +138,7 @@ class KOptListMoveTest {
 
         indexVariableListener.afterListVariableChanged(scoreDirector, e1, 0, 8);
 
-        KOptListMove<TestdataListSolution> kOptListMove = KOptListMove.fromRemovedAndAddedEdges(variableDescriptor,
+        KOptListMove<TestdataListSolution> kOptListMove = fromRemovedAndAddedEdges(variableDescriptor,
                 indexVariableSupply,
                 e1,
                 List.of(
@@ -156,7 +159,7 @@ class KOptListMoveTest {
 
         indexVariableListener.afterListVariableChanged(scoreDirector, e1, 0, 8);
 
-        kOptListMove = KOptListMove.fromRemovedAndAddedEdges(variableDescriptor,
+        kOptListMove = fromRemovedAndAddedEdges(variableDescriptor,
                 indexVariableSupply,
                 e1,
                 List.of(
@@ -172,4 +175,91 @@ class KOptListMoveTest {
         // this move create 2 cycles (v2...v3->t2...) and (v4...v5->v7...v6->v1...v8->v4...)
         assertThat(kOptListMove.isMoveDoable(scoreDirector)).isFalse();
     }
+
+    /**
+     * Create a sequential or non-sequential k-opt from the supplied pairs of undirected removed and added edges.
+     *
+     * @param listVariableDescriptor
+     * @param indexVariableSupply
+     * @param entity The entity
+     * @param removedEdgeList The edges to remove. For each pair {@code (edgePairs[2*i], edgePairs[2*i+1])},
+     *        it must be the case {@code edgePairs[2*i+1]} is either the successor or predecessor of
+     *        {@code edgePairs[2*i]}. Additionally, each edge must belong to the given entity's
+     *        list variable.
+     * @param addedEdgeList The edges to add. Must contain only endpoints specified in the removedEdgeList.
+     * @return A new sequential or non-sequential k-opt move with the specified undirected edges removed and added.
+     * @param <Solution_>
+     */
+    private static <Solution_> KOptListMove<Solution_> fromRemovedAndAddedEdges(
+            ListVariableDescriptor<Solution_> listVariableDescriptor,
+            IndexVariableSupply indexVariableSupply,
+            Object entity,
+            List<Object> removedEdgeList,
+            List<Object> addedEdgeList) {
+
+        if (addedEdgeList.size() != removedEdgeList.size()) {
+            throw new IllegalArgumentException(
+                    "addedEdgeList (" + addedEdgeList + ") and removedEdgeList (" + removedEdgeList + ") have the same size");
+        }
+
+        if ((addedEdgeList.size() % 2) != 0) {
+            throw new IllegalArgumentException(
+                    "addedEdgeList and removedEdgeList are invalid: there is an odd number of endpoints.");
+        }
+
+        if (!addedEdgeList.containsAll(removedEdgeList)) {
+            throw new IllegalArgumentException("addedEdgeList (" + addedEdgeList + ") is invalid; it contains endpoints "
+                    + "that are not included in the removedEdgeList (" + removedEdgeList + ").");
+        }
+
+        Function<Object, Object> successorFunction =
+                getSuccessorFunction(listVariableDescriptor, ignored -> entity, indexVariableSupply);
+
+        for (int i = 0; i < removedEdgeList.size(); i += 2) {
+            if (successorFunction.apply(removedEdgeList.get(i)) != removedEdgeList.get(i + 1)
+                    && successorFunction.apply(removedEdgeList.get(i + 1)) != removedEdgeList.get(i)) {
+                throw new IllegalArgumentException("removedEdgeList (" + removedEdgeList + ") contains an invalid edge ((" +
+                        removedEdgeList.get(i) + ", " + removedEdgeList.get(i + 1) + ")).");
+            }
+        }
+
+        Object[] tourArray = new Object[removedEdgeList.size() + 1];
+        Integer[] incl = new Integer[removedEdgeList.size() + 1];
+        for (int i = 0; i < removedEdgeList.size(); i += 2) {
+            tourArray[i + 1] = removedEdgeList.get(i);
+            tourArray[i + 2] = removedEdgeList.get(i + 1);
+            int addedEdgeIndex = identityIndexOf(addedEdgeList, removedEdgeList.get(i));
+
+            if (addedEdgeIndex % 2 == 0) {
+                incl[i + 1] = identityIndexOf(removedEdgeList, addedEdgeList.get(addedEdgeIndex + 1)) + 1;
+            } else {
+                incl[i + 1] = identityIndexOf(removedEdgeList, addedEdgeList.get(addedEdgeIndex - 1)) + 1;
+            }
+
+            addedEdgeIndex = identityIndexOf(addedEdgeList, removedEdgeList.get(i + 1));
+            if (addedEdgeIndex % 2 == 0) {
+                incl[i + 2] = identityIndexOf(removedEdgeList, addedEdgeList.get(addedEdgeIndex + 1)) + 1;
+            } else {
+                incl[i + 2] = identityIndexOf(removedEdgeList, addedEdgeList.get(addedEdgeIndex - 1)) + 1;
+            }
+        }
+
+        KOptDescriptor<Solution_> descriptor = new KOptDescriptor<>(tourArray,
+                incl,
+                getSuccessorFunction(listVariableDescriptor,
+                        ignored -> entity,
+                        indexVariableSupply),
+                getBetweenPredicate(indexVariableSupply));
+        return descriptor.getKOptListMove(listVariableDescriptor, indexVariableSupply, entity);
+    }
+
+    private static int identityIndexOf(List<Object> sourceList, Object query) {
+        for (int i = 0; i < sourceList.size(); i++) {
+            if (sourceList.get(i) == query) {
+                return i;
+            }
+        }
+        return -1;
+    }
+
 }
